--- intranet/src/core/model/index.ts
@@ -1,11 +1,0 @@
-export interface Restaurant {
-  name: string;
-  address: string;
-  city: string;
-  phoneNumber: string;
-}
-
-export interface Beer {
-  beerId: string;
-  beerName: string;
-}
