+++ intranet/src/core/providers/sign-up/sign-up.model.ts
@@ -1,10 +1,23 @@
-import { Restaurant, Beer } from '#core/model';
+export interface Restaurant {
+  name: string;
+  address: string;
+  city: string;
+  phoneNumber: string;
+}
+
+export interface Beer {
+  id: string;
+  name: string;
+  alcohol: number;
+  volume: number;
+  photoUrl: string;
+}
 
 export interface SignUpModel {
   userName: string;
   password: string;
   email: string;
-  restaurants: Restaurant[];
+  restaurant: Restaurant;
   beers: Beer[];
 }
 
@@ -12,7 +25,12 @@ export const createInitialSignUp = (): SignUpModel => ({
   userName: '',
   password: '',
   email: '',
-  restaurants: [],
+  restaurant: {
+    name: '',
+    address: '',
+    city: '',
+    phoneNumber: '',
+  },
   beers: [],
 });
 
@@ -21,6 +39,6 @@ export interface SignUpContextModel {
   setUserName: (userName: string) => void;
   setPassword: (password: string) => void;
   setEmail: (email: string) => void;
-  setRestaurants: (restaurants: Restaurant[]) => void;
+  setRestaurants: (restaurant: Restaurant) => void;
   setBeers: (beers: Beer[]) => void;
 }
